+++ database/migrations/2026_07_26_122658_create_spare_part_requests_table.php
@@ -17,9 +17,9 @@ return new class extends Migration
             $table->foreignId('client_account_id')->constrained('client_accounts')->cascadeOnDelete();
             // product.id is a plain signed `int` on this app's live schema, same
             // discipline as client_machines.product_id in an earlier migration -
-            // foreignId() would create a bigint unsigned column MySQL refuses as
-            // FK-incompatible with the real product table.
-            $table->unsignedBigInteger('product_id');
+            // foreignId() (and unsignedBigInteger()) would create a column type
+            // MySQL refuses as FK-incompatible with the real product table.
+            $table->integer('product_id');
             $table->unsignedInteger('quantity')->default(1);
             $table->text('note')->nullable();
             $table->enum('status', ['pending', 'approved', 'fulfilled', 'rejected'])->default('pending');
